moved Playlist.tsx to Pages folder
